moved in abstract class the methods to manipulate DB, add users public var in HibernateDBUtilsInMemory
